requirements.txt
Django==4.0
psycopg2-binary==2.9.2
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
pytz==2021.3
requests==2.27.0
boto3==1.21.13
django-storages==1.12.3